pyproject.toml
[project]
name = "library-api"
version = "0.1.0"
description = ""
authors = [{ name = "adheeb2", email = "adheebanvar2@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["alembic (>=1.14.1,<2.0.0)", "fastapi (>=0.115.7,<0.116.0)", "uvicorn (>=0.34.0,<0.35.0)", "sqlalchemy (>=2.0.37,<3.0.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "pydantic-settings (>=2.7.1,<3.0.0)", "python-dotenv (>=1.0.1,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
